cache: use inaugurator cache partition
When inauguratorWithLocalObjectStore is specified rackattack cache is placed unser root partition
in /var/lib/osmosis. When inauguratorWithLocalObjectStore is not specified no cache is used, though inaugurator still
allocated cache lvm partition.

Fix is to use cache LVM partition when inauguratorWithLocalObjectStore is not specified. This is the case when you dont want
osmosis cache inside the product.