inaugurator: add mlx, intel 40G and solarflare nic drivers.
We need those drivers to activate datapath nics inside inaugurator to
provide network health check, and report mac addresses of all nics
to the lab. This is done for several reasons:
1) we want to get an early alert if server is not connected to the switch, in
which case we dont want server to be able to be allocated, but rather issue a report
so infra team can look at the server.
2) we want a live mapping between servers and switch ports that they are connected to.